backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
pydantic[email]==2.8.2
python-dotenv==1.0.1
resend==2.10.0
